chore: Move index related things to the meilidb-core workspace member
