+++ meilidb-core/src/automaton.rs
