+++ meilidb-core/src/data/shared_data.rs
