+++ meilidb-core/src/write_to_bytes.rs
